+++ rocketgram/api/promote_chat_member.py
@@ -24,6 +24,7 @@ class PromoteChatMember(BoolResultMixin, Request):
     can_post_messages: Optional[bool] = None
     can_edit_messages: Optional[bool] = None
     can_delete_messages: Optional[bool] = None
+    can_manage_voice_chats: Optional[bool] = None
     can_invite_users: Optional[bool] = None
     can_restrict_members: Optional[bool] = None
     can_promote_members: Optional[bool] = None